+++ kdecore/date/kcalendarsystemcoptic_p.h
@@ -63,8 +63,8 @@ public:
     virtual bool isProleptic() const;
 
 protected:
-    virtual bool julianDayToDate(int jd, int &year, int &month, int &day) const;
-    virtual bool dateToJulianDay(int year, int month, int day, int &jd) const;
+    virtual bool julianDayToDate(qint64 jd, int &year, int &month, int &day) const;
+    virtual bool dateToJulianDay(int year, int month, int day, qint64 &jd) const;
     KCalendarSystemCoptic(KCalendarSystemCopticPrivate &dd, const KSharedConfig::Ptr, const KLocale *locale);
 
 private:
