+++ kdecore/date/kcalendarsystemislamiccivil_p.h
@@ -69,8 +69,8 @@ public:
     virtual bool isProleptic() const;
 
 protected:
-    virtual bool julianDayToDate(int jd, int &year, int &month, int &day) const;
-    virtual bool dateToJulianDay(int year, int month, int day, int &jd) const;
+    virtual bool julianDayToDate(qint64 jd, int &year, int &month, int &day) const;
+    virtual bool dateToJulianDay(int year, int month, int day, qint64 &jd) const;
     KCalendarSystemIslamicCivil(KCalendarSystemIslamicCivilPrivate &dd, const KSharedConfig::Ptr config, const KLocale *locale);
 
 private:
